Add program 2-strchr.c

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -13,7 +13,7 @@ char *_memcpy(char *dest, char *src, unsigned int n)
 
 	i = 0;
 
-	while (*dest < *(dest + n))
+	while (*dest < *(dest + (int) n))
 	{
 		*(dest + i) = *(src + i);
 		i++;
